KHV-4-reorganisation style
