+++ src/Components/OfferGallery/OfferGallery.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { View, Image, StyleSheet, Text, TouchableOpacity } from "react-native";
+import { View, Image, Text, TouchableOpacity } from "react-native";
 import Swiper from "react-native-swiper";
 import OfferSlides from "../../../Data/OfferSlides";
-
+import BannerStyles from "./OfferGalleryStyles";
 
 const OfferSwiper: React.FC = () => (
   <View style={BannerStyles.container}>
@@ -29,49 +29,5 @@ const OfferSwiper: React.FC = () => (
 const handleSlidePress = (title: string) => {
   console.log(`Slide pressed: ${title}`);
 };
-
-const BannerStyles = StyleSheet.create({
-  container: {
-    marginTop: 50,
-    height: 350,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  slide: {
-    width: 340,
-    height: 300,
-    borderRadius: 6,
-    overflow: "hidden",
-    alignSelf: "center",
-  },
-  image: {
-    width: "100%",
-    height: "100%",
-  },
-  text: {
-    position: "absolute",
-    bottom: 10,
-    left: 10,
-    color: "#fff",
-    fontSize: 14,
-    fontWeight: "bold",
-  },
-  textYear: {
-    position: "absolute",
-    bottom: 10,
-    right: 10,
-    color: "#fff",
-    fontSize: 14,
-    fontWeight: "bold",
-  },
-  textGenre: {
-    position: "absolute",
-    bottom: 10,
-    right: 100,
-    color: "#fff",
-    fontSize: 14,
-    fontWeight: "bold",
-  },
-});
 
 export default OfferSwiper;
