+++ src/Components/TabGallery/BestSellsGallery.tsx
@@ -1,8 +1,9 @@
 import React from "react";
-import { View, Image, Text, TouchableOpacity, StyleSheet } from "react-native";
+import { View, Image, Text, TouchableOpacity } from "react-native";
+import BestSellsStyles from "./TabGalleryStyles";
 
 interface BestSellsGalleryProps {
-  data: { title: string; image: any;}[];
+  data: { title: string; image: any }[];
   onPress: (title: string) => void;
 }
 
@@ -12,54 +13,22 @@ const handleCardPress = (title: string) => {
 
 const BestSellsGallery: React.FC<BestSellsGalleryProps> = ({ data }) => {
   return (
-    <View style={styles.container}>
+    <View style={BestSellsStyles.container}>
       {data.map((item, index) => (
         <TouchableOpacity
           key={index}
-          style={styles.card}
+          style={BestSellsStyles.card}
           onPress={() => handleCardPress(item.title)}>
-          <Image source={item.image} style={styles.image} resizeMode="cover" />
-          <Text style={styles.text}>{item.title}</Text>
+          <Image
+            source={item.image}
+            style={BestSellsStyles.image}
+            resizeMode="cover"
+          />
+          <Text style={BestSellsStyles.text}>{item.title}</Text>
         </TouchableOpacity>
       ))}
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: "row",
-    flexWrap: "wrap",
-    justifyContent: "space-between",
-    paddingHorizontal: 16,
-  },
-  card: {
-    width: "48%",
-    height: 150,
-    borderRadius: 6,
-    overflow: "hidden",
-    marginVertical: 8,
-  },
-  image: {
-    width: "100%",
-    height: "100%",
-  },
-  text: {
-    position: "absolute",
-    bottom: 10,
-    left: 10,
-    color: "#fff",
-    fontSize: 14,
-    fontWeight: "bold",
-  },
-  textSales: {
-    position: "absolute",
-    bottom: 10,
-    right: 10,
-    color: "#fff",
-    fontSize: 14,
-    fontWeight: "bold",
-  },
-});
 
 export default BestSellsGallery;
