+++ src/Components/TabGallery/NewTabGallery.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import { View, Image, Text, TouchableOpacity, StyleSheet } from "react-native";
+import { View, Image, Text, TouchableOpacity } from "react-native";
+import NewTabGalleryStyles from "./NewTabGalleryStyles";
 
 interface NewTabGalleryProps {
   data: { title: string; image: any }[];
@@ -12,46 +13,22 @@ const handleCardPress = (title: string) => {
 
 const NewTabGallery: React.FC<NewTabGalleryProps> = ({ data }) => {
   return (
-    <View style={styles.container}>
+    <View style={NewTabGalleryStyles.container}>
       {data.map((item, index) => (
         <TouchableOpacity
           key={index}
-          style={styles.card}
+          style={NewTabGalleryStyles.card}
           onPress={() => handleCardPress(item.title)}>
-          <Image source={item.image} style={styles.image} resizeMode="cover" />
-          <Text style={styles.text}>{item.title}</Text>
+          <Image
+            source={item.image}
+            style={NewTabGalleryStyles.image}
+            resizeMode="cover"
+          />
+          <Text style={NewTabGalleryStyles.text}>{item.title}</Text>
         </TouchableOpacity>
       ))}
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: "row",
-    flexWrap: "wrap",
-    justifyContent: "space-between",
-    paddingHorizontal: 16,
-  },
-  card: {
-    width: "48%",
-    height: 150,
-    borderRadius: 6,
-    overflow: "hidden",
-    marginVertical: 8,
-  },
-  image: {
-    width: "100%",
-    height: "100%",
-  },
-  text: {
-    position: "absolute",
-    bottom: 10,
-    left: 10,
-    color: "#fff",
-    fontSize: 14,
-    fontWeight: "bold",
-  },
-});
 
 export default NewTabGallery;
